5_illustration.R

+++ Mapping_landscape_ABM/5_illustration.R
@@ -138,7 +138,7 @@ pacmap = import("pacmap")
 set.seed(42)
 
 model = pacmap$PaCMAP(
-  n_components=as.integer(3),
+  n_components=as.integer(2),
   n_neighbors=as.integer(50), 
   MN_ratio=2,
   FP_ratio=10.0,
